# openapi.yaml
openapi: 3.1.0
info:
  title: League Table API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    The backend the League Table frontend (`frontend/services/httpApi.js`) talks to.

    Scoring, the commitments adjustment, ranking and tie-breaking all run on the server;
    the browser only renders the result. A student's raw score is weighted by criterion,
    multiplied by a capped factor from the weekly hours they report and approved by the
    instructor, and clipped at 100. Classmates see only the final adjusted score and rank.

    **Conventions**
    - Every field, query parameter and path parameter is `snake_case`, matching the
      data model in `_docs/spec.md` and FastAPI's defaults. The frontend's services
      layer translates its own JS option names (`weekId`, `windowMode`, ...) at the
      HTTP boundary.
    - Timestamps (`last_updated`, `submitted_at`, ...) are ISO 8601 UTC strings, or
      `null` before any successful read.
    - Errors carry a human-readable `detail` message and an HTTP status; the client
      shows the message to the user.

    **Authentication**
    Every endpoint except `POST /auth/login` and `GET /demo/accounts` needs a session
    cookie. Roles are enforced by the server, never the browser. An `instructor` can do
    everything; a `student` may read the full ranking but only their own breakdown and
    commitments, and may not change settings or approve anything. Sign-in is by access
...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /me/commitments:
    get:
      tags: [Commitments]
      operationId: getMyCommitments
      summary: Your baseline, weekly updates and factor by week (student only)
      responses:
        '200':
          description: The signed-in student's own commitments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyCommitments'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/commitments/baseline:
    put:
      tags: [Commitments]
      operationId: saveMyBaseline
      summary: Submit a semester baseline for approval (student only)
      description: |
        One set of weekly hours for the whole semester. It starts as `pending` and has no
        effect until the instructor approves it. Submitting again while one is pending
        replaces it; submitting after an approval is a baseline change that also waits for
        approval, and the old baseline stays in force meanwhile. Hours go in half-hour steps,
        at most 80 per type and 120 in total.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HoursRequest'
      responses:
        '200':
          description: The pending baseline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Baseline'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /me/commitments/weeks/{week_number}:
    put:
      tags: [Commitments]
      operationId: saveMyWeek
      summary: Set your hours for one week (student only)
      description: |
        Replaces the baseline hours for that week. Only the current and previous week can
        be edited, judged by the Weeks calendar and the server's clock; any other week is
        refused with 403. The update takes effect at once and is logged, but only counts
        once an approved baseline covers that week. A large difference from the baseline is
        flagged in the change log.
      parameters:
        - $ref: '#/components/parameters/WeekNumber'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HoursRequest'
      responses:
        '200':
          description: The weekly update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyUpdate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Commitments]
      operationId: resetMyWeek
      summary: Reset one week to your baseline (student only)
      description: Follows the same edit window. Resetting a week that already follows the baseline does nothing.
      parameters:
        - $ref: '#/components/parameters/WeekNumber'
      responses:
        '204':
          description: The week follows the baseline.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /me/commitments/preview:
    post:
      tags: [Commitments]
      operationId: previewMyAdjustment
      summary: Your factor and adjusted score for hours you have not saved (student only)
      description: |
        For the live preview as a student types. Shows what the hours would give once an
        approved baseline covers the week, using the current settings. Nothing is stored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviewRequest'
      responses:
        '200':
          description: The preview.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /classes/{class_id}/approvals:
    get:
      tags: [Commitments]
      operationId: listApprovals
      summary: Baselines waiting for a decision (instructor only)
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: Pending baselines and baseline changes, with hours by type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approvals'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
  /classes/{class_id}/approvals/{approval_id}:
    post:
      tags: [Commitments]
      operationId: decideApproval
      summary: Approve a baseline from a week, or reject it (instructor only)
      description: |
        Approving makes the baseline count from `effective_week`, by default the current
        week, so earlier weeks keep the factor they had. A baseline it replaces stays in
        force before that week. Rejecting leaves the factor as it was.
      parameters:
        - $ref: '#/components/parameters/ClassId'
        - name: approval_id
          in: path
          required: true
          schema:
            type: string
          example: b3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '200':
          description: The decided baseline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Baseline'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: That baseline has already been decided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
  /classes/{class_id}/students/{student_id}/weeks/{week_number}/reverse:
    post:
      tags: [Commitments]
      operationId: reverseWeeklyUpdate
      summary: Reverse a student's weekly update (instructor only)
      description: |
        Restores the value before the update, whether an earlier update or the baseline,
        and logs the reversal. The student can enter that week again.
      parameters:
        - $ref: '#/components/parameters/ClassId'
        - name: student_id
          in: path
          required: true
          schema:
            type: string
          example: s01
        - $ref: '#/components/parameters/WeekNumber'
      responses:
        '204':
          description: Reversed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Unknown class, student or week, or no weekly update to reverse.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
  /classes/{class_id}/change-log:
    get:
      tags: [Commitments]
      operationId: getChangeLog
      summary: Every commitment change, newest first (instructor only)
      description: |
        Every submission, approval, rejection, weekly update, reset and reversal, and every
        change to the adjustment settings, with who, when, and the old and new values.
      parameters:
        - $ref: '#/components/parameters/ClassId'
        - name: student_id
          in: query
          description: Only this student's entries.
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 200
      responses:
        '200':
          description: Log entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
  /classes/{class_id}/commitments:
    get:
      tags: [Commitments]
      operationId: listStudentCommitments
      summary: Every student's baseline status, hours and factor (instructor only)
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: One row per student.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentCommitments'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly session cookie set by `/auth/login`.
  parameters:
...
    WeekNumber:
      name: week_number
      in: path
      required: true
      description: The week's number in the term (1 to 16), not its ID.
      schema:
        type: integer
      example: 5
...
  responses:
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The caller's role does not allow this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: A student may only open their own breakdown.
    NotFound:
      description: Unknown class.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: |
        A parameter or body field is missing or malformed. Uses FastAPI's default 422
        body, where `detail` is a list of per-field problems.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
...
    HoursRequest:
      type: object
      required: [hours]
      properties:
        hours:
          $ref: '#/components/schemas/Hours'
    BaselineStatus:
      type: string
      enum: [pending, approved, rejected, superseded]
    Baseline:
      type: object
      required: [id, student_id, hours, status, effective_from_week, submitted_at, decided_at]
      properties:
        id:
          type: string
        student_id:
          type: string
        hours:
          $ref: '#/components/schemas/Hours'
        status:
          $ref: '#/components/schemas/BaselineStatus'
        effective_from_week:
          type: [integer, 'null']
          description: The week number it counts from; `null` until approved.
        submitted_at:
          type: string
          format: date-time
        decided_at:
          type: [string, 'null']
          format: date-time
    WeeklyUpdate:
      type: object
      required: [week_id, week_number, hours, entered_at, editable]
      properties:
        week_id:
          type: string
        week_number:
          type: integer
        hours:
          oneOf:
            - $ref: '#/components/schemas/Hours'
            - type: 'null'
          description: '`null` means the student reset this week to their baseline.'
        entered_at:
          type: string
          format: date-time
        editable:
          type: boolean
          description: The week is still inside the edit window.
...
    MyCommitments:
      type: object
      required: [types, limits, baseline, in_effect, weekly_updates, weeks, edit_window]
      properties:
        types:
          type: array
          items:
            $ref: '#/components/schemas/CommitmentType'
        limits:
          $ref: '#/components/schemas/HourLimits'
        baseline:
          oneOf:
            - $ref: '#/components/schemas/Baseline'
            - type: 'null'
          description: The latest submission, whatever its status.
        in_effect:
          oneOf:
            - $ref: '#/components/schemas/Baseline'
            - type: 'null'
          description: The approved baseline that counts now, if any.
        weekly_updates:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyUpdate'
        weeks:
          type: array
          items:
            $ref: '#/components/schemas/WeekFactor'
        edit_window:
          $ref: '#/components/schemas/EditWindow'
    PreviewRequest:
      type: object
      required: [hours]
      properties:
        hours:
          $ref: '#/components/schemas/Hours'
        week_id:
          type: [string, 'null']
          description: Defaults to the student's latest week with scores.
...